(edit) flatten 'shared/SamplesTable' component folder and adjust imports and references
