+++ src/components/DatasetSamplesViewSamplesButton.js
@@ -1,10 +1,10 @@
 import { Box, Heading } from 'grommet'
 import { useModal } from 'hooks/useModal'
 import { useResponsive } from 'hooks/useResponsive'
-import { SamplesTable } from 'components/shared/SamplesTable'
 import { TextCapitalized } from 'components/TextCapitalized'
 import { Button } from 'components/Button'
 import { Modal } from 'components/Modal'
+import { SamplesTable } from 'components/SamplesTable'
 
 export const DatasetSamplesViewSamplesButton = ({
   dataset,
