+++ src/components/ExperimentSamplesTable.js
@@ -3,10 +3,8 @@ import { SamplesContextProvider } from 'contexts/SamplesContext'
 import { useResponsive } from 'hooks/useResponsive'
 import { Column } from 'components/Column'
 import { Row } from 'components/Row'
-import {
-  SamplesTable,
-  SamplesTableAction
-} from 'components/shared/SamplesTable'
+import { SamplesTable } from 'components/SamplesTable'
+import { SamplesTableAction } from 'components/SamplesTableAction'
 
 export const ExperimentSamplesTable = ({ experiment }) => {
   const { setResponsive } = useResponsive()
