+++ src/components/ShowOnlyAddedSamplesFilter.js
