--- src/components/shared/SamplesTable/AccessionCodeCell.js
@@ -1,8 +1,0 @@
-import { memo } from 'react'
-import { TextHighlight } from 'components/TextHighlight'
-
-export const AccessionCodeCell = ({ value }) => {
-  return <TextHighlight>{value}</TextHighlight>
-}
-
-export default memo(AccessionCodeCell)
