--- src/components/shared/SamplesTable/index.js
@@ -1,8 +1,0 @@
-export { SamplesTable } from './SamplesTable'
-export { SamplesTableAction } from './SamplesTableAction'
-export { AdditionalMetadataCell } from './AdditionalMetadataCell'
-export { AccessionCodeCell } from './AccessionCodeCell'
-export { AddRemoveCell } from './AddRemoveCell'
-export { ProcessingInformationCell } from './ProcessingInformationCell'
-export { SampleMetadataCell } from './SampleMetadataCell'
-export { TitleCell } from './TitleCell'
